Add `Part-of:` trailer to patches
I went with `Part-of:` because it feels the most broadly
correct (across forges).

To achieve this, I refactored the loop where it checks
signoffs to also optionally (but by default) add `Part-of:`
trailers.

I think this also fixes a bug where the patches weren't
looped through correctly, so only the first commit was
checked for signoffs.

This also fixes a minor documentation issue, where the git
config key for `pram.repo` was listed as `pram.repository`.

Closes: #7
Signed-off-by: Vivian Heisz (demize) <demize@unstable.systems>