Add tests for `--part-of` and ensure it works locally
This also adds a `--part-of-pr` option to override the
linked PR, specifically for use with local files. I'm not
-thrilled- with this solution, but I tried a few different
ways and this ended up feeling the best: it lets users set
their default to not add `Part-of:` but still use `-p` to
enable it without manually specifying a PR, it allows you
to add the trailer with local patches, and it lets me write
a test for it.

Signed-off-by: Vivian Heisz (demize) <demize@unstable.systems>