Add `-p/--part-of` option to add a trailer linking the PR
This is #9 revived and tweaked a little.

Most of the review feedback should be addressed, and along
with it, things should be optimized a little more. The most
notable change (both in terms of simplified code and in
terms of behavior) is that `linked_pr` is gone, replaced by
`to_close`, per one of the review comments. This has the
effect of adding `Closes:` trailers in some new places,
which I think is generally desirable, as it allows you to
close a pull request while applying it as a patch file or
from an unrelated URL.

Following on from that, `--part-of-pr` has been renamed to
`--link-to`. I've also dropped the `-p` implication here,
as it might be useful for overriding the final `Closes:`
even if the `Part-of:` trailer is unwanted. I've kept the
restriction on only providing one, since I can't see any
justification for more than one here (nor do I understand
how it would -work-).

All tests are also updated to pass with the new trailer,
mostly by disabling it, but a couple pass `--link-to 123`.

Signed-off-by: Vivian Heisz (demize) <demize@unstable.systems>
Signed-off-by: Michał Górny <mgorny@gentoo.org>